Move keyring-related functions to new networking.auth module
